requirements.txt
Django==1.7.4

fann2==1.0.0
Pillow==2.7.0